⚡️(frontend) increase polling interval depend connected users
If we see that someone is connected to the document,
we increase the polling interval. If no one is connected,
we decrease the polling interval.

For the moment, we can only see users connected
with websockets. A good improvement would be to
see users connected with long polling as well.