✨(frontend) create class CollaborationProvider
Create the CollaborationProvider class.
This class is inherited from HocuspocusProvider class.
This class integrate a fallback mechanism to handle the
case where the user cannot connect with websockets.
It will use post request to send the data to the
collaboration server.
It will use an EventSource to receive the data from the
collaboration server.