🐛(backend) compute ancestor_links in get_abilities if needed
The refactor made in the tree view caching the ancestors_links to not
compute them again in the document.get_abilities method lead to a bug.
If the get_abilities method is called without ancestors_links, then they
are computes on all the ancestors but not from the highest readable
ancestor for the current user. We have to compute them with this
constraints and to avoid to compute it again and again we save the
result in a cache.