🐛(frontend) fix callout block arrow navigation
Fixed the behavior of arrow navigation around and between callout
blocks.
When navigating backwards (arrow up or left) from a callout
that is either document's first block or comes right after a callout,
the cursor doesn't go into a text selection, which it should. Similar
behavior if when going forward. This bug is trigged by any of the four
arrow keys.
Added a command to set the cursor position as it is expected.

Signed-off-by: ZouicheOmar <zouiche.dev.omar@gmail.com>